Fixed: disabling curses caused segfaults with /log
This is because when HAVE_CURSES is undefined, phmain.c did not include
intrf.h, which caused td_curses_version() to be undeclared and default
to return type int (but should be const char *). As a result, vprintf
receives an int argument for %s, which coredumps.

This commit changes things so that intrf.h is included unconditionally
(that header has appropriate guards in place already), and moves the
declaration of td_curses_version() outside the guard so that it is
always declared.