Create classifier.sh
I wish to suggest a different system to put recovered files that testdisk users could find useful. Feel free to integrate it in testdisk if you want. Either as a new option for the comand or as a substitute for the current recup_dir1, recup_dir2, recup_dir3 etc system.